+++ src/SensorWrappers/Scd4x.cpp
@@ -2,13 +2,13 @@
 #include "SensirionCore.h"
 #include "Sensirion_UPT_Core.h"
 
-Scd4x::Scd4x(TwoWire& wire) : _wire(wire) {
-    _metaData.deviceType.sensorType = SensorType::SCD4X;
-    _metaData.platform = DevicePlatform::WIRED;
+Scd4x::Scd4x(TwoWire& wire) : mWire(wire) {
+    mMetadata.deviceType.sensorType = SensorType::SCD4X;
+    mMetadata.platform = DevicePlatform::WIRED;
 };
 
 uint16_t Scd4x::start() {
-    _driver.begin(_wire);
+    mDriver.begin(mWire, SCD40_I2C_ADDR_62);
     return 0;
 }
 
@@ -17,7 +17,7 @@ uint16_t Scd4x::measureAndWrite(Measurement measurements[],
     uint16_t co2;
     float temp;
     float humi;
-    uint16_t error = _driver.readMeasurement(co2, temp, humi);
+    uint16_t error = mDriver.readMeasurement(co2, temp, humi);
     if (error) {
         return error;
     }
@@ -25,50 +25,47 @@ uint16_t Scd4x::measureAndWrite(Measurement measurements[],
     measurements[0].signalType = SignalType::CO2_PARTS_PER_MILLION;
     measurements[0].dataPoint.t_offset = timeStamp;
     measurements[0].dataPoint.value = static_cast<float>(co2);
-    measurements[0].metaData = _metaData;
+    measurements[0].metaData = mMetadata;
 
     measurements[1].signalType = SignalType::TEMPERATURE_DEGREES_CELSIUS;
     measurements[1].dataPoint.t_offset = timeStamp;
     measurements[1].dataPoint.value = temp;
-    measurements[1].metaData = _metaData;
+    measurements[1].metaData = mMetadata;
 
     measurements[2].signalType = SignalType::RELATIVE_HUMIDITY_PERCENTAGE;
     measurements[2].dataPoint.t_offset = timeStamp;
     measurements[2].dataPoint.value = humi;
-    measurements[2].metaData = _metaData;
+    measurements[2].metaData = mMetadata;
 
     return HighLevelError::NoError;
 }
 
 uint16_t Scd4x::initializationStep() {
     // stop potentially previously started measurement
-    uint16_t error = _driver.stopPeriodicMeasurement();
+    uint16_t error = mDriver.stopPeriodicMeasurement();
     if (error) {
         return error;
     }
     // Sensor Serial no.
-    uint16_t word0, word1, word2;
-    error = _driver.getSerialNumber(word0, word1, word2);
+    uint64_t serialNumber;
+    error = mDriver.getSerialNumber(serialNumber);
     if (error) {
         return error;
     }
-    uint64_t sensorID = 0;
-    sensorID |=
-        (static_cast<uint64_t>(word0) << 16 * 2) | (word1 << 16) | word2;
 
-    _metaData.deviceID = sensorID;
+    mMetadata.deviceID = serialNumber;
 
     // Start Measurement
-    error = _driver.startPeriodicMeasurement();
+    error = mDriver.startPeriodicMeasurement();
     return error;
 }
 
 SensorType Scd4x::getSensorType() const {
-    return _metaData.deviceType.sensorType;
+    return mMetadata.deviceType.sensorType;
 }
 
 MetaData Scd4x::getMetaData() const {
-    return _metaData;
+    return mMetadata;
 }
 
 size_t Scd4x::getNumberOfDataPoints() const {
@@ -86,5 +83,5 @@ unsigned long Scd4x::getInitializationIntervalMs() const {
 }
 
 void* Scd4x::getDriver() {
-    return reinterpret_cast<void*>(&_driver);
+    return reinterpret_cast<void*>(&mDriver);
 }