+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2]
+
+### Fixed
+- Driver dependecies are now fixed
+- SEN66 and SCD4x serial compilation error fixed
+
 ## [0.3.1]
 
 ### Fixed